Modified Review page, card update actions working

+++ db/migrate/20170223233409_add_grade_to_cards.rb
@@ -1,5 +1,5 @@
 class AddGradeToCards < ActiveRecord::Migration
   def change
-    add_column :cards, :grade, :integer, default: 0
+    add_column :cards, :grade, :float, default: 0
   end
 end
